+++ src/docs/HISTORY.md
@@ -67,6 +67,7 @@ ChangeLog
 - Fixed device name on Windows and Linux which could not map the order in FLTK.
 - Added support for the .heif and .heic movie / picture formats.
 - Fixed monitor device look up on KDE Plasma not finding the right connector (it is different than GNOME).
+- Fixed HDR10+ metadata not getting passed from FFmpeg correctly on each frame.
 - Upgraded OpenJPH to v0.26.0.